Added routing and point rendering

+++ Final Year Project/Assets/Building.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class Building
 {
     public int BuildingNumber {get; set;}
-    public int BuildingHeight {get; set;}
+    public float BuildingHeight {get; set;}
     public List<Vector3> Vertices {get; set;}
 
     public Building()
